@@ -1,4 +1,4 @@
-import { tuft } from '../../src';
+import { tuft } from '../src';
 
 const app = tuft().onError(err => console.error(err));
 
@@ -33,5 +33,5 @@ const server = app.createServer({ port: 3000 });
 server
   .start()
   .then(() => {
-    console.log(`${server.protocol} server listening at http://${server.host}:${server.port}`);
+    console.log(`Server listening at http://${server.host}:${server.port}`);
   });
