@@ -1,4 +1,4 @@
-import { tuft } from '../../src';
+import { tuft } from '../src';
 
 const app = tuft().onError(err => console.error(err));
 
@@ -6,5 +6,5 @@ app.static('/', __dirname)
   .then(async () => {
     const server = app.createServer({ port: 3000 });
     await server.start();
-    console.log(`${server.protocol} server listening at http://${server.host}:${server.port}`);
+    console.log(`Server listening at http://${server.host}:${server.port}`);
   });
