@@ -1,4 +1,4 @@
-import { tuft } from '../../src';
+import { tuft } from '../src';
 
 const app = tuft().onError(err => console.error(err));
 
@@ -20,10 +20,10 @@ app.set('GET /foo/{bar}', t => {
   };
 });
 
-const server = app.createServer({ port: 3000, http1: true });
+const server = app.createServer({ port: 3000 });
 
 server
   .start()
   .then(() => {
-    console.log(`${server.protocol} server listening at http://${server.host}:${server.port}`);
+    console.log(`Server listening at http://${server.host}:${server.port}`);
   });
