@@ -1,36 +1,51 @@
-import { constants } from 'http2';
 import { URLSearchParams } from 'url';
-import { TuftContext, createTuftContext } from '../src/context';
-import {
-  HTTP2_HEADER_STATUS,
-  HTTP2_HEADER_METHOD,
-  HTTP2_HEADER_PATH,
-  HTTP2_METHOD_GET,
-} from '../src/constants';
+import { TuftContext, createTuftContext, TuftRequest } from '../src/context';
+import { HTTP_HEADER_SET_COOKIE } from '../src/constants';
 
-const { HTTP_STATUS_OK } = constants;
+function createMockRequest(method: string = 'GET', url: string = '/') {
+  const mockRequest: any = {
+    method,
+    url,
+    on: jest.fn((_, callback) => {
+      callback();
+    }),
+  };
 
-const mockStream = {
-  on: jest.fn((_, callback) => {
-    callback();
-  }),
-  sentHeaders: {},
-  pushAllowed: false,
-};
+  return mockRequest;
+}
 
-const request = {
-  headers: {},
-  method: 'GET',
-  pathname: '/',
-  searchParams: new URLSearchParams(),
-  params: {},
-  cookies: {},
-  body: null,
-};
+function createMockResponse() {
+  const mockResponse: any = {
+    _headers: {},
+    setHeader: jest.fn((name: string, value: string | number) => {
+      mockResponse._headers[name] = value;
+      return mockResponse;
+    }),
+    getHeader: jest.fn((name: string) => {
+      return mockResponse._headers[name];
+    }),
+    hasHeader: jest.fn((name: string) => {
+      return mockResponse._headers[name] !== undefined;
+    }),
+  };
 
-beforeEach(() => {
-  mockStream.on.mockClear();
-});
+  return mockResponse;
+}
+
+function createMockTuftRequest(method: string = 'GET', pathname: string = '/') {
+  const tuftRequest: TuftRequest = {
+    headers: {},
+    method,
+    pathname,
+    search: '',
+    searchParams: new URLSearchParams(),
+    params: {},
+    cookies: {},
+    body: null,
+  };
+
+  return tuftRequest;
+}
 
 /**
  * TuftContext
@@ -39,70 +54,89 @@ beforeEach(() => {
 describe('TuftContext', () => {
   describe('new TuftContext()', () => {
     test('returns an instance of TuftContext with the expected properties', () => {
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
+      const mockTuftRequest = createMockTuftRequest();
+
       const t = new TuftContext(
-        //@ts-expect-error
-        mockStream,
-        request,
+        mockRequest,
+        mockResponse,
+        mockTuftRequest,
       );
 
       expect(t).toBeInstanceOf(TuftContext);
-      expect(t).toHaveProperty('request', request);
-      expect(t).toHaveProperty('outgoingHeaders', {});
+      expect(t).toHaveProperty('request', mockTuftRequest);
     });
   });
 
   describe('TuftContext.prototype.setHeader()', () => {
-    test('returns undefined', () => {
+    test('returns the instance of TuftContext', () => {
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
+      const mockTuftRequest = createMockTuftRequest();
+
       const t = new TuftContext(
-        //@ts-expect-error
-        mockStream,
-        request,
+        mockRequest,
+        mockResponse,
+        mockTuftRequest,
       );
 
-      expect(t.setHeader(HTTP2_HEADER_STATUS, HTTP_STATUS_OK)).toBe(t);
+      expect(t.setHeader('custom-header-name', 'custom-header-value')).toBe(t);
     });
   });
 
   describe('TuftContext.prototype.getHeader()', () => {
     test('returns the expected value', () => {
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
+      const mockTuftRequest = createMockTuftRequest();
+
       const t = new TuftContext(
-        //@ts-expect-error
-        mockStream,
-        request,
+        mockRequest,
+        mockResponse,
+        mockTuftRequest,
       );
 
-      t.setHeader(HTTP2_HEADER_STATUS, HTTP_STATUS_OK);
+      t.setHeader('custom-header-name', 'custom-header-value');
 
-      expect(t.getHeader(HTTP2_HEADER_STATUS)).toBe(HTTP_STATUS_OK);
+      expect(t.getHeader('custom-header-name')).toBe('custom-header-value');
     });
   });
 
   describe('TuftContext.prototype.setCookie()', () => {
+    const mockRequest = createMockRequest();
+    const mockResponse = createMockResponse();
+    const mockTuftRequest = createMockTuftRequest();
+
     const t = new TuftContext(
-      //@ts-expect-error
-      mockStream,
-      request,
+      mockRequest,
+      mockResponse,
+      mockTuftRequest,
     );
 
     test('adds `set-cookie` to the `outgoingHeaders` property', () => {
-      expect(t.outgoingHeaders).not.toHaveProperty('set-cookie');
+      expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toBeUndefined();
       expect(t.setCookie('a', 'foo')).toBe(t);
-      expect(t.outgoingHeaders).toHaveProperty('set-cookie', ['a=foo; Path=/']);
+      expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toEqual(['a=foo; Path=/']);
     });
 
-    test('updates the `set-cookie` entry of the `outgoingHeaders` property', () => {
-      expect(t.outgoingHeaders).toHaveProperty('set-cookie');
-      expect(t.setCookie('b', 'foo')).toBe(t);
-      expect(t.outgoingHeaders).toHaveProperty('set-cookie', ['a=foo; Path=/', 'b=foo; Path=/']);
+    test('updates `set-cookie` on the `outgoingHeaders` property', () => {
+      expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toBeDefined();
+      expect(t.setCookie('b', 'bar')).toBe(t);
+      expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toEqual(['a=foo; Path=/', 'b=bar; Path=/']);
     });
   });
 
   describe('TuftContext.prototype.setCookie()', () => {
     describe('when passed an options object', () => {
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
+      const mockTuftRequest = createMockTuftRequest();
+
       const t = new TuftContext(
-        //@ts-expect-error
-        mockStream,
-        request,
+        mockRequest,
+        mockResponse,
+        mockTuftRequest,
       );
 
       test('adds the expected cookie entry', () => {
@@ -117,30 +151,38 @@ describe('TuftContext', () => {
           httpOnly: true,
         });
 
-        expect(t.outgoingHeaders).toHaveProperty('set-cookie', [
+        expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toEqual([
           `a=foo; Expires=${expires.toUTCString()}; Max-Age=1000; Domain=example.com; Path=/; Secure; HttpOnly`,
         ]);
       });
     });
 
     describe('when passed an options object with an invalid property', () => {
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
+      const mockTuftRequest = createMockTuftRequest();
+
       const t = new TuftContext(
-        //@ts-expect-error
-        mockStream,
-        request,
+        mockRequest,
+        mockResponse,
+        mockTuftRequest,
       );
 
       test('adds the expected cookie entry', () => {
         t.setCookie('a', 'foo', { invalidProperty: 42 });
-        expect(t.outgoingHeaders).toHaveProperty('set-cookie', ['a=foo; Path=/']);
+        expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toEqual(['a=foo; Path=/']);
       });
     });
 
     describe('when passed an options object with `secure` and `httpOnly` set to false', () => {
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
+      const mockTuftRequest = createMockTuftRequest();
+
       const t = new TuftContext(
-        //@ts-expect-error
-        mockStream,
-        request,
+        mockRequest,
+        mockResponse,
+        mockTuftRequest,
       );
 
       test('adds the expected cookie entry', () => {
@@ -149,60 +191,76 @@ describe('TuftContext', () => {
           httpOnly: false,
         });
 
-        expect(t.outgoingHeaders).toHaveProperty('set-cookie', ['a=foo; Path=/']);
+        expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toEqual(['a=foo; Path=/']);
       });
     });
 
     describe('when passed an options object with `sameSite` set to `strict`', () => {
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
+      const mockTuftRequest = createMockTuftRequest();
+
       const t = new TuftContext(
-        //@ts-expect-error
-        mockStream,
-        request,
+        mockRequest,
+        mockResponse,
+        mockTuftRequest,
       );
 
       test('adds the expected cookie entry', () => {
         t.setCookie('a', 'foo', { sameSite: 'Strict' });
-        expect(t.outgoingHeaders).toHaveProperty('set-cookie', ['a=foo; SameSite=Strict; Path=/']);
+        expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toEqual(['a=foo; SameSite=Strict; Path=/']);
       });
     });
 
     describe('when passed an options object with `sameSite` set to `lax`', () => {
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
+      const mockTuftRequest = createMockTuftRequest();
+
       const t = new TuftContext(
-        //@ts-expect-error
-        mockStream,
-        request,
+        mockRequest,
+        mockResponse,
+        mockTuftRequest,
       );
 
       test('adds the expected cookie entry', () => {
         t.setCookie('a', 'foo', { sameSite: 'Lax' });
-        expect(t.outgoingHeaders).toHaveProperty('set-cookie', ['a=foo; SameSite=Lax; Path=/']);
+        expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toEqual(['a=foo; SameSite=Lax; Path=/']);
       });
     });
 
     describe('when passed an options object with `sameSite` set to `none`', () => {
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
+      const mockTuftRequest = createMockTuftRequest();
+
       const t = new TuftContext(
-        //@ts-expect-error
-        mockStream,
-        request,
+        mockRequest,
+        mockResponse,
+        mockTuftRequest,
       );
 
       test('adds the expected cookie entry', () => {
         t.setCookie('a', 'foo', { sameSite: 'None' });
-        expect(t.outgoingHeaders).toHaveProperty('set-cookie', ['a=foo; SameSite=None; Path=/']);
+        expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toEqual(['a=foo; SameSite=None; Path=/']);
       });
     });
 
     describe('when passed an options object with `sameSite` set to an invalid value', () => {
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
+      const mockTuftRequest = createMockTuftRequest();
+
       const t = new TuftContext(
-        //@ts-expect-error
-        mockStream,
-        request,
+        mockRequest,
+        mockResponse,
+        mockTuftRequest,
       );
 
       test('adds the expected cookie entry', () => {
         //@ts-expect-error
         t.setCookie('a', 'foo', { sameSite: 'foo' });
-        expect(t.outgoingHeaders).toHaveProperty('set-cookie', ['a=foo; Path=/']);
+        expect(t.getHeader(HTTP_HEADER_SET_COOKIE)).toEqual(['a=foo; Path=/']);
       });
     });
   });
@@ -215,15 +273,12 @@ describe('TuftContext', () => {
 describe('createTuftContext()', () => {
   describe('with no options', () => {
     test('returns an instance of TuftContext', () => {
-      const headers = {
-        [HTTP2_HEADER_METHOD]: HTTP2_METHOD_GET,
-        [HTTP2_HEADER_PATH]: '/',
-      };
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
 
       const result = createTuftContext(
-        //@ts-expect-error
-        mockStream,
-        headers,
+        mockRequest,
+        mockResponse,
       );
 
       expect(result).toBeInstanceOf(TuftContext);
@@ -232,15 +287,12 @@ describe('createTuftContext()', () => {
 
   describe('with a path that includes a query string', () => {
     test('returns an instance of TuftContext', () => {
-      const headers = {
-        [HTTP2_HEADER_METHOD]: HTTP2_METHOD_GET,
-        [HTTP2_HEADER_PATH]: '/foo?bar=baz',
-      };
+      const mockRequest = createMockRequest('GET', '/foo?bar=baz');
+      const mockResponse = createMockResponse();
 
       const result = createTuftContext(
-        //@ts-expect-error
-        mockStream,
-        headers,
+        mockRequest,
+        mockResponse,
       );
 
       expect(result).toBeInstanceOf(TuftContext);
@@ -249,11 +301,8 @@ describe('createTuftContext()', () => {
 
   describe('with option `params` set', () => {
     test('returns an instance of TuftContext', () => {
-      const headers = {
-        [HTTP2_HEADER_METHOD]: HTTP2_METHOD_GET,
-        [HTTP2_HEADER_PATH]: '/foo/bar/baz',
-      };
-
+      const mockRequest = createMockRequest('GET', '/foo/bar/baz');
+      const mockResponse = createMockResponse();
       const options = {
         params: {
           0: 'one',
@@ -262,9 +311,8 @@ describe('createTuftContext()', () => {
       };
 
       const result = createTuftContext(
-        //@ts-expect-error
-        mockStream,
-        headers,
+        mockRequest,
+        mockResponse,
         options,
       );
 
